use multisig address as url param
